+++ examples/complete-alb/versions.tf
@@ -1,14 +1,10 @@
 terraform {
-  required_version = ">= 0.13.1"
+  required_version = ">= 1.0"
 
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 3.40"
-    }
-    random = {
-      source  = "hashicorp/random"
-      version = ">= 2.0"
+      version = ">= 4.59"
     }
     null = {
       source  = "hashicorp/null"